Support custom working-directory (#3)
Previously, the orb hardcoded the clone location to ~/project. If someone wants to use /mnt/ramdisk or some other dir instead, this would break their build. We now add the parameter with a sane default.

As best as I can tell, CircleCI does not natively expose the `working_directory` value to the pipeline params.